pyproject.toml
[tool.poetry]
name = "rcmt-packages"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"ruamel.yaml" = "^0.17.17"

[tool.poetry.dev-dependencies]
rcmt = "~0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
